rename menu component to offer component
