+++ frontend/frontend/src/app/app-routing.module.ts
@@ -1,22 +1,21 @@
 import { UserReservationsComponent } from './user-reservations/user-reservations.component';
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { LoginComponent } from './login/login.component';
 import { RegisterComponent } from './register/register.component';
 import { HomeComponent } from './home/home.component';
-import { MenuComponent } from './menu/menu.component';
-import { FoodOfferComponent } from './menu/food-offer/food-offer.component';
+import { OfferComponent } from './offer/offer.component';
+import { FoodOfferComponent } from './offer/food-offer/food-offer.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent, pathMatch: 'full' },
   { path: 'login', component: LoginComponent },
   { path: 'register', component: RegisterComponent },
   { path: 'home', component: HomeComponent },
-  { path: 'offer', component: MenuComponent },
+  { path: 'offer', component: OfferComponent },
   { path: 'my-reservations', component: UserReservationsComponent },
   { path: 'menu', component: FoodOfferComponent },
-  { path: 'games', component: MenuComponent },
+  { path: 'games', component: OfferComponent },
 ];
 
 @NgModule({
