+++ frontend/frontend/src/app/offer/food-offer/food-offer.component.ts
