+++ frontend/frontend/src/app/offer/offer-nav-bar/offer-nav-bar.component.spec.ts
