+++ frontend/frontend/src/app/offer/offer.component.html
